hosts: docker/sftp: Improve sleepfd() implementation
It was introduced with reliable sleepx() as bash(1) alternative and used
for standard input redirection and infinite sleep for all interpreters.

It works well, but has few places for improvement:

  1) mksleepfd() reader job has poor naming in job list breaking logs:
      [1]   Terminated              while IFS=''; do
          read _ || :;
          done 0<> "$fl"
  2) sleep_inf() has duplicated code with reader job (see (1)).
  3) Two global variables $sleepjob and $sleepfl with only $sleepjob
     being validated gives more room for misuse
  4) Open coded standard input redirection with $sleepfl variable
  5) busy-loop waiting for reader job to come up with valid $sleepfl

Introduce sleepfd() that is defined when calling mksleepfd(). Use it
from reader job and sleep_inf() to fix naming issue in (1) and code
duplication in (2).

Introduce sleepfd_redirect_stdin() and use it in sleepfd() and globally.
Hide $sleepfl within it and place more strict checks on $sleepjob when
preparing $sleepfl path. Make sure redirect target is a named pipe that
is not current standard input before redirecting addressing (3) and (4).

Busy-loop in (5) can be eliminated by redirecting standard input for
whole `if' statement that does not spawn subshell, opens standard input
file descriptor that inherited by sleepfd() job and thus known to be
valid. Filesystem entry for named pipe removed right after redirection
reducing time for race.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>