hosts: docker/sftp: Use bash(1) as fallback when standard input isn't terminal
It was observed with nginx running under pm2 supervisor that dash(1)
forces job control to off because process isn't attached to terminal:

   dash: 1: set: can't access tty; job control turned off

However dash(1) does not treat this as fatal error and continues
execution with job control turned off that in turn leads to misbehavior.

This issue isn't the case when it runs as docker entrypoint, with pid 1
and terminal attached (i.e. docker-run(1) --tty=true option).

To support both use cases with single codebase "test -t 0" if standard
input is a terminal device and run main() using bash(1), that does not
seem to depend on terminal device for job control.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>